move merging logic to dataset cache
