+++ qcodes/dataset/sqlite/queries.py
@@ -1796,109 +1796,27 @@ def remove_trigger(conn: ConnectionPlus, trigger_id: str) -> None:
     transaction(conn, f"DROP TRIGGER IF EXISTS {trigger_id};")
 
 
-def load_new_data_from_db_and_append(
-            conn: ConnectionPlus,
-            table_name: str,
-            rundescriber: RunDescriber,
-            write_status: Dict[str, Optional[int]],
-            read_status: Dict[str, int],
-            existing_data: Mapping[str, Mapping[str, np.ndarray]],
-    ) -> Tuple[Dict[str, Optional[int]],
-               Dict[str, int],
-               Dict[str, Dict[str, np.ndarray]]]:
-    """
-    Append any new data in the db to an already existing datadict and return the merged
-    data.
-
-    Args:
-        conn: The connection to the sqlite database
-        table_name: The name of the table the data is stored in
-        rundescriber: The rundescriber that describes the run
-        write_status: Mapping from dependent parameter name to number of rows
-          written to the cache previously.
-        read_status: Mapping from dependent parameter name to number of rows
-          read from the db previously.
-        existing_data: Mapping from dependent parameter name to mapping
-          from parameter name to numpy arrays that the data should be
-          inserted into.
-          appended to.
-
-    Returns:
-        Updated write and read status, and the updated ``data``
-
-    """
-    new_data, updated_read_status = _load_new_data(
-        conn, table_name, rundescriber, read_status
-    )
-
-    (updated_write_status,
-     merged_data) = append_shaped_parameter_data_to_existing_arrays(
-        rundescriber,
-        write_status,
-        existing_data,
-        new_data
-    )
-    return updated_write_status, updated_read_status, merged_data
-
-
-def append_shaped_parameter_data_to_existing_arrays(
-        rundescriber: RunDescriber,
-        write_status: Dict[str, Optional[int]],
-        existing_data: Mapping[str, Mapping[str, np.ndarray]],
-        new_data: Mapping[str, Mapping[str, np.ndarray]],
-) -> Tuple[Dict[str, Optional[int]],
-           Dict[str, Dict[str, np.ndarray]]]:
-    """
-    Append datadict to an already existing datadict and return the merged
-    data.
-
-    Args:
-        rundescriber: The rundescriber that describes the run
-        write_status: Mapping from dependent parameter name to number of rows
-          written to the cache previously.
-        new_data: Mapping from dependent parameter name to mapping
-          from parameter name to numpy arrays that the data should be
-          appended to.
-        existing_data: Mapping from dependent parameter name to mapping
-          from parameter name to numpy arrays of new data.
-
-    Returns:
-        Updated write and read status, and the updated ``data``
-    """
-    parameters = tuple(ps.name for ps in
-                       rundescriber.interdeps.non_dependencies)
-    merged_data = {}
-
-    updated_write_status = copy(write_status)
-
-    for meas_parameter in parameters:
-
-        existing_data_1_tree = existing_data.get(meas_parameter, {})
-
-        new_data_1_tree = new_data.get(meas_parameter, {})
-
-        shapes = rundescriber.shapes
-        if shapes is not None:
-            shape = shapes.get(meas_parameter, None)
-        else:
-            shape = None
-
-        (merged_data[meas_parameter],
-         updated_write_status[meas_parameter]) = _merge_data(
-            existing_data_1_tree,
-            new_data_1_tree,
-            shape,
-            single_tree_write_status=write_status.get(meas_parameter)
-        )
-    return updated_write_status, merged_data
-
-
-def _load_new_data(
+def load_new_data_for_rundescriber(
         conn: ConnectionPlus,
         table_name: str,
         rundescriber: RunDescriber,
         read_status: Mapping[str, int],
 ) -> Tuple[Dict[str, Dict[str, np.ndarray]], Dict[str, int]]:
+    """
+    Load all new data for a given rundesciber since the rows given by read_status.
+
+    Args:
+        conn: The connection to the sqlite database
+        table_name: The name of the table the data is stored in
+        rundescriber: The rundescriber that describes the run
+        read_status: Mapping from dependent parameter name to number of rows
+          read from the db previously.
+
+    Returns:
+        new data and an updated number of rows read.
+
+    """
+
     parameters = tuple(ps.name for ps in
                        rundescriber.interdeps.non_dependencies)
     updated_read_status: Dict[str, int] = dict(read_status)
@@ -1918,81 +1836,3 @@ def _load_new_data(
         new_data_dict[meas_parameter] = new_data
         updated_read_status[meas_parameter] = start + n_rows_read - 1
     return new_data_dict, updated_read_status
-
-
-def _merge_data(existing_data: Mapping[str, np.ndarray],
-                new_data: Mapping[str, np.ndarray],
-                shape: Optional[Tuple[int, ...]],
-                single_tree_write_status: Optional[int]
-                ) -> Tuple[Dict[str, np.ndarray], Optional[int]]:
-
-    subtree_merged_data = {}
-    subtree_parameters = set(existing_data.keys()) | set(new_data.keys())
-    new_write_status: Optional[int] = None
-    for subtree_param in subtree_parameters:
-        existing_values = existing_data.get(subtree_param)
-        new_values = new_data.get(subtree_param)
-        if existing_values is not None and new_values is not None:
-            (subtree_merged_data[subtree_param],
-             new_write_status) = _insert_into_data_dict(
-                existing_values,
-                new_values,
-                single_tree_write_status,
-                shape=shape
-            )
-        elif new_values is not None:
-            (subtree_merged_data[subtree_param],
-             new_write_status) = _create_new_data_dict(
-                new_values,
-                shape
-            )
-        elif existing_values is not None:
-            subtree_merged_data[subtree_param] = existing_values
-            new_write_status = single_tree_write_status
-
-    return subtree_merged_data, new_write_status
-
-
-def _create_new_data_dict(new_values: np.ndarray,
-                          shape: Optional[Tuple[int, ...]]
-                          ) -> Tuple[np.ndarray, int]:
-    if shape is None:
-        return new_values, new_values.size
-    else:
-        n_values = new_values.size
-        data = np.zeros(shape, dtype=new_values.dtype)
-
-        if new_values.dtype.kind == "f" or new_values.dtype.kind == "c":
-            data[:] = np.nan
-
-        data.ravel()[0:n_values] = new_values.ravel()
-        return data, n_values
-
-
-def _insert_into_data_dict(
-        existing_values: np.ndarray,
-        new_values: np.ndarray,
-        write_status: Optional[int],
-        shape: Optional[Tuple[int, ...]]
-) -> Tuple[np.ndarray, Optional[int]]:
-    if shape is None or write_status is None:
-        return np.append(existing_values, new_values, axis=0), None
-    else:
-        if existing_values.dtype.kind in ('U', 'S'):
-            # string type arrays may be too small for the new data
-            # read so rescale if needed.
-            if new_values.dtype.itemsize > existing_values.dtype.itemsize:
-                existing_values = existing_values.astype(new_values.dtype)
-        n_values = new_values.size
-        new_write_status = write_status+n_values
-        if new_write_status > existing_values.size:
-            log.warning(f"Incorrect shape of dataset: Dataset is expected to "
-                        f"contain {existing_values.size} points but trying to "
-                        f"add an amount of data that makes it contain {new_write_status} points. Cache will "
-                        f"be flattened into a 1D array")
-            return (np.append(existing_values.flatten(),
-                              new_values.flatten(), axis=0),
-                    new_write_status)
-        else:
-            existing_values.ravel()[write_status:new_write_status] = new_values.ravel()
-            return existing_values, new_write_status
